third_party/WebKit/LayoutTests/fast/history/visited-link-button.html

<script src="../../resources/js-test.js"></script>

<style>
input {
    line-height: 10px;
    display: none;
}
a:visited input {
    background-color: yellow;
}
</style>

<a id="a" href=""><input id="input" type="button"></input></a>

<script>
if (window.testRunner)
    testRunner.keepWebHistory();

a.style.color = "orange";
a.style.color = "";
shouldBeEqualToString("getComputedStyle(input).lineHeight", "normal");
</script>
