third_party/blink/web_tests/platform/generic/fast/history/visited-link-button-expected.txt

PASS getComputedStyle(input).lineHeight is "normal"
PASS successfullyParsed is true

TEST COMPLETE
